src/backend/requirements.txt
agentops==0.2.6
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multion==1.3.5
packaging==23.2
pluggy==1.5.0
psutil==5.9.8
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.1
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
termcolor==2.4.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
